// src/events/dto/price.dto.ts
export class PriceDto {
  EditableMinPrice: number;

  Enabled: boolean;

  IsBase: boolean;

  IsBest: boolean;

  IsEditable: boolean;

  IsEditableForWeb: boolean;

  LayerTypeId: null;

  MinPrice: number;

  ModeOfSaleOfferId: number;

  Offer: boolean;

  PackageId: number;

  ParentPackageId: number;

  PerformanceId: number;

  PerformancePriceTypeId: number;

  Price: number;

  PriceTypeId: number;

  ZoneId: number;
}