// src/events/dto/seat.dto.ts
export class SeatDto {
  AisleIndicator: string;

  AllocationId: number;

  DisplayLetter: string;

  HasStairs: boolean;

  HoldCodeId: number;

  Id: number;

  IsSeat: boolean;

  LogicalSeatNumber: number;

  LogicalSeatRow: number;

  ScreenId: number;

  SeatNumber: string;

  SeatRow: string;

  SeatStatusId: number;

  SeatTypeId: number;

  SectionId: number;

  XPosition: number;

  YPosition: number;

  ZoneId: number;
}