// src/events/dto/section.dto.ts
export class SectionDto {
  AdditionalText: string;

  AdditionalText2: string;

  CreateLocation: null | string;

  CreatedBy: null | string;

  CreatedDateTime: null | string;

  Description: string;

  Id: number;

  PrintDesc: string;

  PrintSequence: number;

  SectionLegend: string;

  ShortDesc: string;

  UpdatedBy: string;

  UpdatedDateTime: string;
}